Add ProfilePictureUrl property to Instructor and Student DTOs
Added a new property `ProfilePictureUrl` of type `string` with a
default value of an empty string to the `InstructorReadDTO` class
in the `EdufyAPI.DTOs.InstructorDTOs` namespace. This property
will store the URL of the instructor's profile picture.

Added a new property `ProfilePictureUrl` of type `string` with a
default value of an empty string to the `GetStudentsDTO` class
in the `EdufyAPI.DTOs.StudentDTOs` namespace. This property
will store the URL of the student's profile picture.